Store the cap-table abi value in the .MIPS.abiflags isa_ext field
This value is currently unused so we can use it to store which cap-table
ABI is currently being used. This is useful in order to prevent mixing
object files linked with incompatible ABIs.

It should currently be possible to mix in many cases but it may break in
surprising ways so it is better to just insist on a consistent ABI choice
across all .o files.

See CTSRD-CHERI/lld#20