Stop clearing $c13 when not necessary at -O2 or higher
$c13 is used for on-stack function arguments. Previously we were always
setting it to null when calling a function without on-stack arguments.
However, it turns out this results in noticeable overhead (over 2.5% of
isntructions in libc.so were cgetnull $c13). When compiling at -O2 or
higher we no longer clear $c13 if we are calling a function without stack
args from a function without stack args (i.e. $c13 was already null on
function entry). This means we now only need to clear $c13 when we are in
a function with more than 8 args or a varargs function are calling a
function with less than 9 args.