Adjust tests to $c13 clearing change
Seems like in two of the tests that are affected the delay slot can't be
filled sensibly but in the other it can. It looks better in the broken lld
tests where quite a few functions are now smaller (I should really stop
writing lld tests in C since they break on every compiler change).